feat(ui): alert create/edit form basic reworking
